+++ app/routes/settings/account/index.tsx
@@ -17,7 +17,7 @@ import { Button } from "~/components/ui/button";
 import { Form, FormControl, FormField, FormItem, FormMessage } from "~/components/ui/form";
 import { Input } from "~/components/ui/input";
 
-import WarningComponent from "~/components/warning";
+import WarningComponent from "~/components/warning-dialog";
 
 import { ChevronRight, ExternalLink, TriangleAlert, type LucideIcon } from "lucide-react";
